requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.7.0
openai==1.45.0
python-telegram-bot==21.1.1
requests~=2.31.0
tenacity==8.3.0
wolframalpha~=5.0.0
duckduckgo_search==5.3.1b1
spotipy~=2.23.0
pytube~=15.0.0
gtts~=2.5.1
whois~=0.9.27
Pillow~=10.3.0
httpx==0.27.0
readability-lxml==0.8.1
lxml-html-clean==0.1.1
youtube-transcript-api==0.6.2
asyncpg==0.29.0
google-api-python-client==2.140.0
